Ports tg's cavegen, bitfields, area_flags, improved random mineral rocks (#4800)
* Refactor bitfields global, improve turret_flags debugging (#53239)

* Refactors area stuff (#52751)

-bitfielded a bunch of bools on /area, I left some untouched cus they get called a lot
-Unused vars
-Fixed a var pretending to be a fake bool
-Probably more

* Random mineral spawns change less turfs. (#49205)

As-is, random mineral spawning changes the turf to the subtype with that mineral, which results in a lot of turf changes I guess. Now it just changes the mineralType of the turf. Unless it's gibtonite or a cave spawner, because the alternatives are worse and those aren't all that common anyway.

🆑
tweak: Random mineral turfs no longer cause large amounts of turf changes on initialization.
/🆑

* Refactor the procedural generation of lavaland and turf/closed/mineral (#54915)

This replaces lavaland's old diagonal tunnel gen which was really
horrendously jammed into asteroid floor code (?????) with Cellular
Automata which runs in rust (PR for that here:
https://github.com/tgstation/rust-g/pull/57 ). The new code is a bit
cleaner, but also looks better.

VID: https://streamable.com/a45ke2

Things to do:
- Make an icemoon version
- Fix the roundstart atmos adjacency issues

I asked AnturK if this was an acceptable PR for this month; he said it
was okay as long as I didn't add new areas, which I don't plan to do.
But if anyone thinks this PR breaks the spirit of the month I'll open
it again in december.

* fix (#55462)

* Fix lavaland map

* Fix bad merge

Co-authored-by: Urumasi <Urumasi@email.cz>

Co-authored-by: Jared-Fogle <35135081+Jared-Fogle@users.noreply.github.com>
Co-authored-by: TiviPlus <57223640+TiviPlus@users.noreply.github.com>
Co-authored-by: Buggy123 <buggy1997123@yahoo.com>
Co-authored-by: Qustinnus <Floydje123@hotmail.com>
Co-authored-by: Urumasi <Urumasi@email.cz>