Modernises uplinks (#4851)
* initial

* a

* nah

* Change uplink implants to inherit the uplink flags of the uplink they were bought from (#59735)

Uplinks implants bought from uplinks inherit the uplink flags of the uplink they were bought from.

This fixes an undocumented bug where by purchasing an uplink implant, nuke ops would be able to get things they normally can't.

Co-authored-by: tralezab <40974010+tralezab@users.noreply.github.com>
Co-authored-by: Jordan Brown <Cyberboss@users.noreply.github.com>

* Fixes pop restricted uplink items not properly unlocking (#58542)

Makes it so if you become a traitor when specific items are pop locked, they will update accordingly as the population increases.

* [s] Fixes uplink discount exploit. (#58625)

* epic

* split em

* early return

* whatever

* Add "Implant" to list of uplink locations (#51828)

* 👛 Add "Implant" to list of uplink locations

:cl: coiax
add: You may now pick Implant as a preferred uplink location, meaning
that if you become a traitor, you will start with an uplink implant. The
cost of the implant (4 TC) is deducted from your total starting TC,
because the Syndicate doesn't give freebies.
fix: Admins removing an uplink from a person with an uplink implant will
now destroy the implant, rather than leaving an implant that does
nothing.
/:cl:

If you are the sort of person who gets their PDA, pen and radio stolen
really easily, then maybe you'd be interested in just spawning with an
implant. Be warned, all of the 17-20 TC items will be out of reach
because of the implant cost.

Could be considered slightly more powerful than just ordering the
implant at shift start, given you don't have to get rid of an implanter
and suspicious box anymore.

When testing with the admin grant uplink buttons, I discovered that the
"Take" uplink option was only half working when taking the uplink from a
person with only an implant uplink. It would take the component, but not
the implant, leaving the person with a do-nothing action button. This
has been fixed.

* Add missing DEFINES

* Remove one layer of spaces

Will this sate the linting gods?

* Replace 20 default with TELECRYSTALS_DEFAULT

* default

* former price

* Warn player about cost of implant uplink (#58872)

Co-authored-by: celotajstg <celotajstg@users.noreply.github.com>

* missing preference

* Fixes a roundstart uplink implanting runtime (#60242)

There's no guarantee there will be a user making the implanting, such as it's the case of roundstart equipping.

* prefs

* Fixes uplink type being unchangeable and radio uplink frequency (#58908)

* Fixes uplink-related bugs

* Actually fixes radio uplink properly

* all good

* uplink ui name

* undocumented stuff removed

* resolves conflict

* fix

Co-authored-by: tralezab <40974010+tralezab@users.noreply.github.com>
Co-authored-by: Jordan Brown <Cyberboss@users.noreply.github.com>
Co-authored-by: Seth Scherer <supernovaa41@protonmail.com>
Co-authored-by: celotajstg <celotajstg@users.noreply.github.com>
Co-authored-by: Rohesie <rohesie@gmail.com>
Co-authored-by: Unknown <faceaccont@gmail.com>