Give EVA and External airlock access to the CMO (#5908)
* Give EVA and External airlock access to the CMO

EMT has access to these doors but the CMO does not, And CMO has access to the shutters of EVA but not EVA itself.

If there is no EMT the CMO ought to be able to retrieve close-ranged spaced people.
(Yes the CMO has it's own suit but it lacks access to a jetpack and large oxygen tanks)

* Remove ACCESS_APOTHECARY from my pull

no more ACCESS_APOTHECARY as of https://github.com/BeeStation/BeeStation-Hornet/pull/5917