Ports Mail From /tg/ which was ported from Kiwi [TM Ready] (#6263)
* You got mail!

And it's terribly broken

* You got Spam!

But no actual mail...

* You ACTUALLY got mail

IT works! (but it's not over, yet...)

* Everyone got mail!

And now onto the other stuff to do...

* Small fixes...

* You got Lint!

* You got mail (and the spam is invisible)

* Lint momento

* You got mail! (and now the spam is visible)

also Junkmail.txt in wip

* You got commit

Just to avoid overwriting...

* You got Mail! (it's not a mess)

And this time, it's ready for testing

* spare fixes and unused stuff removal

* Admin logging in case critical mail has recieved

* You got new mailman suit!

also Plasmaman suit

* further update

* Lazarus if Ian is die

* You got updates!

Many changes related to sprites, also new spritebag

* the mailbag can now hold milk and bread

blame the coderbus discord for this

* oop

* Fixed something related to plasmaman.dm

and also brought back the new hat in the sprites

* Double whoops

now envirohemlets should make sense

* FINALLY should fix the mail envirohelmets

* Fixed broken clothing

Blame the previous commit

* first batch of fixes

* 2nd batch

* Big batch of fixes pointed by archainal + free renaming of stuff

* forgot about dest_tagger

* Fixed more stuff

also the mail list now on it's own dedicated file

* new hats are back again

* Last minute fixes

Gotta hurry for the Testmerge!

* Fixed wrappers

* fixes for the time being

* shuttle_areas momento

* forgor to check

skull emoji

* there we go

* fixed stuff and merge update

* forgor the envirosuits

* Fixes codewise and icon wise and whatnot

no clue why the item pools still mix

* Centcom, not Centcomm

* momento do mergefix

* You got Patches!

Fixed the spawners being weird finally and the nuke code is now fixed

* small code fixup

* Map mergefix

* removed the last trace of the debtor, removed unused code

* sum more fixes

* Finally fixed the damn job list bug

Thanks Bacon and MNarath!

* 1Z corg now works with the mail PR