Corg Bureacracy Update (+ Robo Ghost Sniffer) (#7527)
* Corg Bureacracy Update

Modifies the front-facing cargo lobby to add a desk so cargo cargo can interact with the crew without needing to open the airlock.

Add additional chest drawers for storing paperwork in the cargo office and HoP's office.

Tweaks position of RD's and CMO's stamps to not be buried underneath other items at round start, removes the Captain's duplicate stamp in their quarters, adds the HoS's stamp to their office, and adds a spare set of "Approved" and "Denied" stamps to the Quartermaster's office (similar to MetaStation)

* Adds Ecto Sniffer to Robotics

Adds Ecto Sniffer to Robotics to bring it in-line with other maps currently in rotation.